Binary mirrors for solvers: CUDD, CVC4 and Z3.
This is meant to avoid re-compiling the solvers each time during CI.
The env variable PYSMT_BIN_MIRROR can be set to define the location of
the binary mirror. Note that this is not implemented for all solvers.
Finally, CVC4 required some slightly different handling, thus the
pythonpath in shippable.yml has been changed.