SMTLIB: Parser Refactoring
Refactoring of the parser to allow extensibility

Simplified extensibility of the parser by splitting the
special-purpose code in the main loop in separate function invoked
using a map.

A few minor hacks to gain performance (~2 seconds executing all the tests in tests/smtlib)

Added suport for escaping in SmtLib printer and parser. It is now
possible to work with symbols having arbitrary names except for SmtLib
keywords. Differently from SmtLib standard, we allow escaping in symbols:
|here is an example of \| inside symbol|