Issue #220: Logic as String
Top level functions from Factory (and thus most shortcuts) can now
take a logic as a string. Instead of writing

  is_sat(f, logic=pysmt.logics.QF_LRA)

it is now possible to write:

  is_sat(f, logic="QF_LRA")

This is handled in Factory._get_solver_class(). Other code within
pySMT should always work under the assumption that logics are given as
a pysmt.logics object.