Refactoring Solver options handling:
Solver constructor (and shortuct) takes **options as free keyword
arguments. These options are checked by the class SolverOptions, in
order to validate that these are meaningful options and perform a
preliminary validation to catch typos etc. by raising a ValueError
exception if the option is unknown.

It is now possible to do:

  Solver(name="bdd", dynamic_reordering=True)

NOTE: By using the construct above, one assumes that the chosen solver
      supports the given option. While this should usually the case,
      this approach does not allow to gracefully ignore unkown options.
      An alternative approach, could be to issue a warning, instead of
      raising an exception.

This commit changes the constructor of Solver by

- Removing the argument user_options
- Introducing the named arguments **options