Strings Theory: Examples and walkers extension
Introduce two examples in example.py using Strings. This stimulates a
reasonably big part of pySMT, and highlighted missing pieces in the
printing/parsing and oracles.

All those parts have been extended accordingly.

NOTE: This commit changes the notation for HR printing of quoted
symbols:

- 'str str' is a Symbol with name str<space>str
- "str str" is a String (constant) with value str<space>str

NOTE: We change the behavior of Simplifier.validate_simplification.
Instead of being a Boolean value, it becomes the name of a Solver to be
used to validate the simplification. If set to None, no simplification
is performed.