BUGFIX: Add delay after starting Wrapper to detect dead process
If the process of the generic SMT-Wrapper did not start yet, the write
on the pipe is queued. If the process dies (or never starts) this does
not result in an IOError, but it looks like an empty read from the pipe.
To avoid this, we way 0.01s after creating the process, to make sure
that the write will either raise an IOError or be successful.