install.py: Cmd-line args override env variable
If both PYSMT_SOLVER env variable is set and a command line argument is
given, install.py will ignore the env variable.

 $ export PYSMT_SOLVER="z3"
 $ ./install.py --msat

will install MathSAT and ignore z3.

The idea is that the command line is more specific and should have
priority.

This is also needed to allow the installation of a single solver in the
CI when using the "all" setting.