Z3: Extend back conversion to create new Symbols
Using Z3 Tactics, it is possible to create expressions in which new
variables are introduced. The Z3Converter was using the get_symbols
method to force the usage of existing symbols. This has now been
extended, and new Symbols are created on the fly if needed.

For now only a warning is given, but we should consider whether this
behavior needs a special option to be disabled.