Z3: Avoid AstRef usage within Z3Converter
Intermediate Z3 objects constructed during the forward conversion are
not wrapped anymore by an AstRef object. Only the last object (the one
returned by Z3Converter.convert) is a proper AstRef object.  In this way
we save time and memory (20%) when performing the conversion, because we
avoid creating many useless intermediate objects. At the same time,
object obtained from convert() are still full-fledged Z3 objects, and
can be used with the native Z3 API.

In order to achieve this, we need to make ther Ref counting explicit in
the Converter. The ref counter is increased when the expression is
created, then it is cached, and does not need to be increased
anymore. During the destruction of the Converter, all ref counters will
be decreased. In this way, Z3 objects created within the Z3Converter *can*
survive the Z3Converter distruction.