Enforce a single call to is_sat if not incremental
Non incremental is_sat modifies the stack of assertions. For this
reason, it should be possible to call it only once.

This achieves this by raising an error in solve() if called after a call
to is_sat.

This also clarifies the documentation of Solver.get_model.

Thanks to @colinmorris for the feedback on Solver interface.