Removed option 'quantified' from Solver
As a side-effect, writing Solver('msat') now leads to the expected
result of obtaining msat.

In order to force a particular type of logic, the keyword 'logic' must
be used. To identify the correct logic, it is possible to use an oracle,
or obtain the non-quantified logic and do 'logic.get_quantified_version()'