TypeManager: New Environment object
The TypeManager is a factory components for all the Types.

For compatibility, base types are still defined in pysmt.typing, but
this should make it easier to be uniform in the handling of types.

We extend the Normalization process (FormulaManager.normalize) in order
to handle also type information.

TODO: Maybe we need to consider also walk_function?