CVC4: Force absolute import of CVC4
The python bindings name clashes with the modules names both in
solvers and installers. This can create problems in python 2 on some
systems.

We address this by requiring the use of absolute import. Note that we
already do the same for z3.

Thanks to Alexey Ignatiev (@2sev) for bringin this issue to our
attention.