TypesOracle and sort declaration in SMT-LIB script
Custom types were not being declared within the SMT-LIB
scripts. Therefore, serializing an expression containing custom types
would generate an SMT-LIB file that could not be parsed.

To address this issue, we introduce a new Oracle: TypesOracle. The
main method of the oracle is ``get_types``, that provides the complete
list of types occuring in an expression. The result of this method is
a partially sorted list, in which simpler types appear before
composite types in which the former types are included. See
test_oracles.py:test_types_oracle for more information.