Logics: Add custom_types in Theory
Any formula that uses a custom type will belong to the theory with
custom_types=True. By default all SMT-LIB logics are defined without
custom_types (False). This is due to the fact that most of the SMT-LIB
benchmarks do not require the possibility of declaring types, and not
all solvers (e.g., Boolector) support this.