CVC4: Free C++ SmtEngine when Solver is destroyed
The SWIG wrapper is not handling properly the ownership of either
SmtEngine or ExprManager, resulting in a memory leak: When the python
objects are release the underlying C++ are not destroyed.

This commit works-around the issue by forcing the ownership of the
SmtEngine to the python wrapper. When the wrapper is destroyed, the C++
is also destroyed.

A proper patch for CVC4 should be designed in order to properly account
also for the release of the ExprManager, and the fact that ExprManager
is ref'd within SmtEngine (but SWIG has no way of knowing this).