BUG: Detect BV in Array and Function in infix op
In the expression Select(s, i).Equals(5) we need to understand the
resulting type of the Select in order to cast the 5 correctly.  However,
this operation was performed using a custom function in fnode.py, that
was missing several cases, e.g., the same problem existed for Function
applications.

The fix involves using get_type().is_bv_type() instead.

This commit also cleans-up the error checking related to infix
operation, by moving the check of whether it is enabled into a
decorator, and providing a more meaningful error.