BV: Convert GMPY object in BV constructor
Automatically convert GMPY object in the BV constructor.

Previously, BV(mpz(5), 8) would create a node with a mpz object in the
payload *independently* of whether GMPY was enabled or not.

The behavior of the BV constructor is now consisten with the behavior of
Int and Real.

Tests for this have been added in test_constants.py