Fix StopIteration use in SmtLib parser (#521)
Introduced two different functions for consuming a token in the parser: consume_maybe and consume. The first will consume a new token if available, but stop the iteration if no token is available. The second will require that a token is available and throw an exception if this is not the case.

This makes the intent in the code much more clear. Previously, we were relying on StopIteration bubbling up the call stack to obtain the behavior of consume_maybe. However, this would not allow us to distinguish between a stream that ended abruptly and one that ended normally. 

This also fixes a bug in the management of extra_queue in the parser