Fixed CI
Fixed an issue introduced in a recent release of Azure Pipelines. It seems that the python running on the CI machines does not correctly identify the library path where libpython is with the LIBDIR config variable. This commit exposes an environment varible that allows the use to specift the python library path by hand.