CLA: Add utility to check for CLA
We require all contributors to submit a change to CONTRIBUTORS as a
way to confirming acceptance of the CLA.

To check this, we compare the git log with the entries in the
CONTRIBUTORS file.

Since users might have sent commits with different names/emails, we
introduce a .mailmap to provide canonical names.

The CONTRIBUTORS data has been updated with the content of CLAHub.