Remove six and python 2 compatibility code
We officially dropped support for python 2 in 0.9.0 (2020-04-26).
We now remove the dependency from six and all its usages in the code.
This means that pySMT now does not work anymore with python 2.

This change is motivated by python 2 EOL and aimed at reducing
maintenance overhead.

Thanks to the six.py team for the awesome work!