Parse logic str in Portfolio
Portfolio does not rely on the factory infrastructure, but instead we
instantiate the class directly. In the factory module we take care of
converting the logic from str to a pysmt.logic object.
This step was missing in the Portfolio constructor.

We could do the conversion later, but I think it makes sense to ensure
that logic is always a pysmt object.

Fixes #708.